requirements.txt
Django==5.2
django-cas-ng==5.1.0
python-dotenv==1.2.2
msal==1.35.1